Extract `jacobi` module; add `Neg` impl
Factors `JacobiSymbol` out of the `traits` module and into its own
`jacobi` module.

Also adds a `Neg` impl, adapted from the one in `crypto-primes`, which
seems to be the one thing missing for `crypto-primes` to switch to it
for its Jacobi symbol type.
